Split out the code for parsing an index file from DirCache module.
